moves connectedMember types to file
