+++ src/store/connectedMember/types.ts
@@ -1,0 +1,5 @@
+export type ConnectedMemberState = {
+  delegateKey: string;
+  isActiveMember: boolean;
+  memberAddress: string;
+} | null;
